#####1.优化CATextField在安卓端的体验

#####2.优化CATextView在安卓端的体验

#####3.优化CAWebView在安卓端的体验

#####4.增加对安卓以及iOS屏幕旋转支持

#####5.修复CABar设置TitleView布局不正确的bug

#####6.修改CADevice中加速度计的更新间隔

#####7.修改CADevice中消息通知在安卓平台的延迟时间

#####8.其他一些细节的优化

+++ CrossApp/proj.android/src/org/CrossApp/lib/CrossAppGLSurfaceView.java
@@ -58,7 +58,7 @@ public class CrossAppGLSurfaceView extends GLSurfaceView {
 		return mGLSurfaceView;
 	}
 
-	public static void queueAccelerometer(final float x, final float y, final float z, final long timestamp) 
+	public static void queueAccelerometer(final float x, final float y, final float z, final float timestamp)
 	{
 		mGLSurfaceView.queueEvent(new Runnable()
 		{
